数据库表详情页新增大小字段

+++ src/main/java/com/changhong/opendb/driver/Table.java
@@ -36,7 +36,7 @@ public class Table
         /**
          * 数据库大小（KB）
          */
-        private Integer size;
+        private Float size;
 
         /**
          * 数据行数
